requirements.txt
# Telegram Bot API
python-telegram-bot>=21.0
